🎨 Refactoring demo app
- Move code to componnents
- Move files to demo/src folder
