+++ demo/index.js
@@ -1,7 +1,7 @@
 /* @flow */
 import { AppRegistry } from 'react-native';
 
-import App from './App';
+import App from './src/App';
 import { name as appName } from './app.json';
 
 AppRegistry.registerComponent(appName, () => App);
